pyproject.toml
[project]
name = "multi-agent-system"
version = "0.1.0"
description = "Multi Agent System multi-agent system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "jsonschema>=4.19,<5",
  "PyYAML>=6.0,<7",
  "pytest>=8.0,<9",
  "typer>=0.12,<1",
  "pydantic>=2,<3",
  "python-dotenv>=1,<2",
  "httpx>=0.27,<0.28",
  "pytest-asyncio>=0.23,<0.24",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
